Move internal under schema package, the place of its use.
Also rename 'protoparser' to 'parser' since the qualifier is no longer useful in an environment which only deal with protos.
