+++ wire-schema/src/main/java/com/squareup/wire/schema/EnumType.java
@@ -19,7 +19,7 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.LinkedHashMultimap;
 import com.google.common.collect.Multimap;
 import com.squareup.wire.WireType;
-import com.squareup.wire.internal.protoparser.EnumElement;
+import com.squareup.wire.schema.internal.parser.EnumElement;
 import java.util.Collection;
 import java.util.Map;
 import java.util.NavigableSet;
