+++ wire-schema/src/main/java/com/squareup/wire/schema/Extend.java
@@ -17,8 +17,8 @@ package com.squareup.wire.schema;
 
 import com.google.common.collect.ImmutableList;
 import com.squareup.wire.WireType;
-import com.squareup.wire.internal.protoparser.ExtendElement;
-import com.squareup.wire.internal.protoparser.FieldElement;
+import com.squareup.wire.schema.internal.parser.ExtendElement;
+import com.squareup.wire.schema.internal.parser.FieldElement;
 
 public final class Extend {
   private final ExtendElement element;
