+++ wire-schema/src/main/java/com/squareup/wire/schema/Rpc.java
@@ -16,7 +16,7 @@
 package com.squareup.wire.schema;
 
 import com.squareup.wire.WireType;
-import com.squareup.wire.internal.protoparser.RpcElement;
+import com.squareup.wire.schema.internal.parser.RpcElement;
 
 public final class Rpc {
   private final RpcElement element;
