+++ wire-schema/src/main/java/com/squareup/wire/schema/internal/parser/ExtensionsElement.java
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.squareup.wire.internal.protoparser;
+package com.squareup.wire.schema.internal.parser;
 
 import com.google.auto.value.AutoValue;
-import com.squareup.wire.internal.Util;
+import com.squareup.wire.schema.internal.Util;
 import com.squareup.wire.schema.Location;
 
-import static com.squareup.wire.internal.Util.appendDocumentation;
+import static com.squareup.wire.schema.internal.Util.appendDocumentation;
 
 @AutoValue
 public abstract class ExtensionsElement {
