+++ wire-schema/src/main/java/com/squareup/wire/schema/internal/parser/Nullable.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.squareup.wire.internal.protoparser;
+package com.squareup.wire.schema.internal.parser;
 
 import java.lang.annotation.Retention;
