+++ wire-schema/src/test/java/com/squareup/wire/schema/SchemaTest.java
@@ -16,7 +16,7 @@
 package com.squareup.wire.schema;
 
 import com.squareup.wire.WireType;
-import com.squareup.wire.internal.Util;
+import com.squareup.wire.schema.internal.Util;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
