+++ wire-schema/src/test/java/com/squareup/wire/schema/internal/parser/EnumElementTest.java
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.squareup.wire.internal.protoparser;
+package com.squareup.wire.schema.internal.parser;
 
 import com.google.common.collect.ImmutableList;
 import com.squareup.wire.schema.Location;
 import org.junit.Test;
 
-import static com.squareup.wire.internal.protoparser.OptionElement.Kind.STRING;
+import static com.squareup.wire.schema.internal.parser.OptionElement.Kind.STRING;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public final class EnumElementTest {
