+++ wire-schema/src/test/java/com/squareup/wire/schema/internal/parser/OptionElementTest.java
@@ -13,17 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.squareup.wire.internal.protoparser;
+package com.squareup.wire.schema.internal.parser;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import org.junit.Test;
 
-import static com.squareup.wire.internal.protoparser.OptionElement.Kind.BOOLEAN;
-import static com.squareup.wire.internal.protoparser.OptionElement.Kind.LIST;
-import static com.squareup.wire.internal.protoparser.OptionElement.Kind.MAP;
-import static com.squareup.wire.internal.protoparser.OptionElement.Kind.OPTION;
-import static com.squareup.wire.internal.protoparser.OptionElement.Kind.STRING;
+import static com.squareup.wire.schema.internal.parser.OptionElement.Kind.BOOLEAN;
+import static com.squareup.wire.schema.internal.parser.OptionElement.Kind.LIST;
+import static com.squareup.wire.schema.internal.parser.OptionElement.Kind.MAP;
+import static com.squareup.wire.schema.internal.parser.OptionElement.Kind.OPTION;
+import static com.squareup.wire.schema.internal.parser.OptionElement.Kind.STRING;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public final class OptionElementTest {
