+++ wire-schema/src/test/java/com/squareup/wire/schema/internal/parser/ParsingTester.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.squareup.wire.internal.protoparser;
+package com.squareup.wire.schema.internal.parser;
 
 import com.squareup.wire.schema.Location;
 import java.io.File;
