+++ wire-schema/src/test/java/com/squareup/wire/schema/internal/parser/ProtoParserTest.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.squareup.wire.internal.protoparser;
+package com.squareup.wire.schema.internal.parser;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.squareup.wire.internal.Util;
-import com.squareup.wire.internal.protoparser.OptionElement.Kind;
+import com.squareup.wire.schema.internal.Util;
+import com.squareup.wire.schema.internal.parser.OptionElement.Kind;
 import com.squareup.wire.schema.Location;
 import com.squareup.wire.schema.ProtoFile;
 import java.util.Arrays;
